requirements.txt
customtkinter==5.2.0
requests==2.31.0
psutil==5.9.0
ping3==4.0.4
speedtest-cli==2.1.3